Bugfix | GUI/UI fixes | And some additions | v0.1.3.2 Added CanvasLayer components to UI scenes. Added Game manager for better scenetree control. (Not applied)
Expiremental: Trying to get a signle value for all directions of a
vector2, I'm almost there, but adding the (x,y) values doubles the
output when moving towards a corner, a problem I'm trying to solve.

Fixed player melee not working. (Fogot to setup it up with the collsion
system)
Fixed UI components being affected by lighting. (I think)

Note: I also removed a collider and ajusted the melee box a bit.
Note: I'm changing the way the scenes are handled using a root node, but
I haven't applied this to the main system as it's in development.

Optimized code.