Level Designer | MODDING! | v0.2.1
This update brings us closer to a working mod system.
In the past few days, I worked up some programming energy and have been
attempting to program and design a level builder! Now it has a custom
tilemap loader, and a custom map builder with save and loading of
previous saves! There is so much planned for this so please stay tuned!

I know I have taken my focus of the base game, but that is temporary as
this system is going to be used by us devs to make the main game as
well! A massive learning expirence for me!!! I'm proud of my
acomplishments and research so far.

Note: Some things like the save directory will be exluded from the core
game upon release and documentation will be formed for modding.

Note: This is only the level designer, because I have been mostly
focused on it, I've had no time to incorperate a load system from the
menu.

Changes

Added save system
Added load system
Added Keybinds for loading
Cleaned up some code.

Added controls view for the designer and other components for the mdk.

A lot more that my head hurts too bad to explain, look at the code.