Level Designer | Tilesets | Modding | Texture support improvements | v0.2.5
Added camera to pan with in the level designer.
Added verbosity to the tileset loader.
Added create_tile function for devs.

Improved tilesets to load and index more effectively!

Changed the way the selector moved in order to fix a bug.

Fixed a bug for loading and saving tilesets.

This commit fixes a lot of bugs and re-introduces tilesets.

A word from me.

I'm Kilo, the indie developer of this game that maintains this
open-source project. I just want to say that modding can be done now,
the development tool is actually a really easy to access one, but it
will be up to the modders to handle their resource files aside from the
already read-only files accessable from the game at runtime. Simply
create a built-in script inside of your mod project. You need to do it
directly from your root scene of a level exported with the Level
Designer Tool.
NOTE: If you have scripts or any resource in your root sceen when saving
from the Level Designer, it will be wiped from your root sceen, if you
follow the path of modding, only edit your level from inside of Godot. A
documentation will be made for this in the future.