Tweaks | MDKTool | Modding Changes | v0.2.5.2
Restructured files/folders
Added MDKTool (NON-FUNCTIONL) yet ;)
There's fixes and other sorts of things, I did remove the map editor button because I'm integrating it with the MDKTool feature, but it'll be back in a future update.
Another thing I'm adding is the OBUS or the On-Board Updating System. It's also non-function yet. These were features I was tweaking out and testing. The MDK is recent, but I've been working on the OBUS since December. Scripts did change, but I can't really make them out because I restructured the whole game.