Porting changes, not functional or official, early stage! [Project on HOLD]
Final push before halting development temporarily (TBA return). These are some modifications to the project for Godot 4.0 they are all still being worked on and the game is planned to be reworked entirely from the ground up at some point or another project will be started, this all depends on resources available at that time.